club/serializers/club_apply_serializers.py
from rest_framework import serializers

from club.models import ClubApply


class ClubApplySerializer(serializers.ModelSerializer):
    class Meta:
        model = ClubApply
        fields = "__all__"
