pyproject.toml
[tool.poetry]
name = "checkin"
version = "0.1.0"
description = ""
authors = ["devch1013 <devch1013@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.4"
djangorestframework = "^3.15.2"
pre-commit = "^4.0.1"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.0.1"
djangorestframework-simplejwt = "^5.3.1"
ruff = "^0.8.4"
pytest-django = "^4.9.0"
pytest-cov = "^6.0.0"
django-redis = "^5.4.0"
loguru = "^0.7.3"
django-debug-toolbar = "^4.4.6"
django-extensions = "^3.2.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
